cmd/internal/obj/loong64, cmd/compile, runtime: move the loong64 G to R31
R22 is traditionally the frame pointer on loong64, so while we aren't
necessarily using it like so, it's probably better to move G elsewhere.
Moving it to the end of the GPR sequence seems cleaner in that it makes
the reserved GPRs more contiguous.

Breakage is not expected, given that no code should rely on the actual
register number of g, and the port is still relatively young.

Change-Id: I2e03b736f61567f3321a479da991e7ca552f3e73