cmd/link: bump function alignment on loong64 to 16 bytes
The LA464 micro-architecture is particularly sensitive to alignment of
hot loops, with performance fluctuations of around ±10% possible, as
can be seen on some microbenchmarks (e.g. FmtFprintfPrefixedInt,
RegexpMatchEasy0_1K or Revcomp in the go1 benchmark suite) due to
uncontrolled code layout sending critical loop heads to underaligned
positions.

In addition to this effect, the PCALIGN directive works with PCs
relative to beginning of functions. So if the function alignment is
less than that requested by PCALIGN, the following code may in fact not
be aligned as such, leading to unexpected performance.

The current function alignment on loong64 is 8 bytes, which seems to
stem from mips64 or riscv64. In order to make performance more
predictable on loong64, it is raised to 16 bytes to ensure that at
least `PCALIGN $16` works.

As alignment of loops written in Go is yet to be tackled, and the
codegen is not otherwise touched, benchmark numbers for this change are
not going to be meaningful, and not included.

Change-Id: I2120ef3746ce067e274920c82091810073bfa3be