internal/cpu, runtime: make linux/loong64 HWCAP data available
This can be used to toggle runtime usages of ISA extensions as such
usages appear.

Only the CRC32 bit is exposed for now, as the others are not going to be
utilized in the standard library for a while.

Change-Id: I774032ca84dc8bcf1c9f17558917315af07c7314
(cherry picked from commit 109d9c3806e0e8dcfddba820144e040b51afb425)