cmd/link: ignore symbols prefixed with "wm{4,8}." in .group sections
These symbols can show up in libgcc.a with debug info, which will
interfere with internal linking if not ignored like a few others.
See GCC commit 520cda8ca8971319 for details regarding production of
such symbols; they do not seem to be arch-specific, although test
failures are observed on a linux/loong64 system in my case.

Change-Id: I6e1ced3665e4915a97af050e6e2b393e082a3a14